Moved some code & RNG Scaffolding
