+++ .gitignore
@@ -5,3 +5,4 @@ obj/
 riderModule.iml
 /_ReSharper.Caches/
 .idea/
+.vs/
